[icon-button-unstyled] Split IconButton into unstyled and styled button
